+++ yellowstone-grpc-proto/src/plugin/filter/filter.rs
@@ -1185,6 +1185,7 @@ mod tests {
             commitment: None,
             accounts_data_slice: Vec::new(),
             ping: None,
+            from_slot: None,
         };
         let limit = FilterLimits::default();
         let filter = Filter::new(&config, &limit, &mut create_filter_names());
@@ -1216,6 +1217,7 @@ mod tests {
             commitment: None,
             accounts_data_slice: Vec::new(),
             ping: None,
+            from_slot: None,
         };
         let mut limit = FilterLimits::default();
         limit.accounts.any = false;
@@ -1251,6 +1253,7 @@ mod tests {
             commitment: None,
             accounts_data_slice: Vec::new(),
             ping: None,
+            from_slot: None,
         };
         let mut limit = FilterLimits::default();
         limit.transactions.any = false;
@@ -1285,6 +1288,7 @@ mod tests {
             commitment: None,
             accounts_data_slice: Vec::new(),
             ping: None,
+            from_slot: None,
         };
         let mut limit = FilterLimits::default();
         limit.transactions.any = false;
@@ -1325,6 +1329,7 @@ mod tests {
             commitment: None,
             accounts_data_slice: Vec::new(),
             ping: None,
+            from_slot: None,
         };
         let limit = FilterLimits::default();
         let filter = Filter::new(&config, &limit, &mut create_filter_names()).unwrap();
@@ -1389,6 +1394,7 @@ mod tests {
             commitment: None,
             accounts_data_slice: Vec::new(),
             ping: None,
+            from_slot: None,
         };
         let limit = FilterLimits::default();
         let filter = Filter::new(&config, &limit, &mut create_filter_names()).unwrap();
@@ -1453,6 +1459,7 @@ mod tests {
             commitment: None,
             accounts_data_slice: Vec::new(),
             ping: None,
+            from_slot: None,
         };
         let limit = FilterLimits::default();
         let filter = Filter::new(&config, &limit, &mut create_filter_names()).unwrap();
@@ -1503,6 +1510,7 @@ mod tests {
             commitment: None,
             accounts_data_slice: Vec::new(),
             ping: None,
+            from_slot: None,
         };
         let limit = FilterLimits::default();
         let filter = Filter::new(&config, &limit, &mut create_filter_names()).unwrap();
@@ -1575,6 +1583,7 @@ mod tests {
             commitment: None,
             accounts_data_slice: Vec::new(),
             ping: None,
+            from_slot: None,
         };
         let limit = FilterLimits::default();
         let filter = Filter::new(&config, &limit, &mut create_filter_names()).unwrap();